+++ resources/views/tailwind/includes/table.blade.php
@@ -33,7 +33,12 @@
                 <x-livewire-tables::table.cell :colspan="count($bulkActions) ? count($columns) + 1 : count($columns)">
                     @unless ($selectAll)
                         <div>
-                            <span>{!! __('You have selected <strong>:count</strong> rows, do you want to select all <strong>:total</strong>?', ['count' => $rows->count(), 'total' => number_format($rows->total())]) !!}</span>
+                            <span>
+                                @lang('You have selected')
+                                <strong>{{ $rows->count() }}</strong>
+                                @lang('rows, do you want to select all')
+                                <strong>{{ number_format($rows->total()) }}</strong>?
+                            </span>
 
                             <button
                                 wire:click="selectAll"
@@ -45,7 +50,11 @@
                         </div>
                     @else
                         <div>
-                            <span>{!! __('You are currently selecting all <strong>:total</strong> rows.', ['total' => number_format($rows->total())]) !!}</span>
+                            <span>
+                                @lang('You are currently selecting all')
+                                <strong>{{ number_format($rows->total()) }}</strong>
+                                @lang('rows').
+                            </span>
 
                             <button
                                 wire:click="resetBulk"
@@ -72,7 +81,7 @@
                         <div class="flex rounded-md shadow-sm">
                             <input
                                 wire:model="selected"
-                                value="{{ $row->getKey() }}"
+                                value="{{ $row->{$primaryKey} }}"
                                 onclick="event.stopPropagation();return true;"
                                 type="checkbox"
                                 class="rounded-md shadow-sm border-gray-300 block transition duration-150 ease-in-out sm:text-sm sm:leading-5"
@@ -91,7 +100,7 @@
                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M20 13V6a2 2 0 00-2-2H6a2 2 0 00-2 2v7m16 0v5a2 2 0 01-2 2H6a2 2 0 01-2-2v-5m16 0h-2.586a1 1 0 00-.707.293l-2.414 2.414a1 1 0 01-.707.293h-3.172a1 1 0 01-.707-.293l-2.414-2.414A1 1 0 006.586 13H4" />
                         </svg>
 
-                        <span class="font-medium py-8 text-gray-400 text-xl">@lang('No items found. Try narrowing your search.')</span>
+                        <span class="font-medium py-8 text-gray-400 text-xl">@lang($emptyMessage)</span>
                     </div>
                 </x-livewire-tables::table.cell>
             </x-livewire-tables::table.row>